requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
djangorestframework==3.13.1
importlib-metadata==4.11.4
Markdown==3.3.7
pytz==2022.1
sqlparse==0.4.2
zipp==3.8.0
